+++ tests/test_connector_sharded.py
@@ -16,43 +16,59 @@ import os
 import sys
 import time
 
+from pymongo import MongoClient, ReadPreference
+
 sys.path[0:0] = [""]
 
-from mongo_connector.connector import Connector
+from mongo_connector.connector import Connector, get_mininum_mongodb_version
 from mongo_connector.doc_managers.doc_manager_simulator import DocManager
 from mongo_connector.test_utils import (assert_soon,
                                         db_user,
                                         db_password,
                                         ShardedClusterSingle)
+from mongo_connector.version import Version
+
 from tests import unittest, SkipTest
 
 
-class TestConnectorSharded(unittest.TestCase):
+class ShardedConnectorTestCase(unittest.TestCase):
 
-    def setUp(self):
+    @classmethod
+    def setUpClass(cls):
         if db_user and db_password:
             auth_args = dict(auth_username=db_user, auth_key=db_password)
         else:
             auth_args = {}
-        self.cluster = ShardedClusterSingle().start()
-        self.dm = DocManager()
-        self.connector = Connector(
-            mongo_address=self.cluster.uri,
-            doc_managers=[self.dm],
+        cls.cluster = ShardedClusterSingle().start()
+        cls.main_uri = cls.cluster.uri + '/?readPreference=primaryPreferred'
+        cls.dm = DocManager()
+        cls.connector = Connector(
+            mongo_address=cls.main_uri,
+            doc_managers=[cls.dm],
             **auth_args
         )
-        self.connector.start()
+        cls.connector.start()
+        assert_soon(lambda: len(cls.connector.shard_set) == 2,
+                    message='connector failed to find both shards!')
 
-    def tearDown(self):
-        self.connector.join()
+    @classmethod
+    def tearDownClass(cls):
+        cls.connector.join()
         try:
             os.unlink('oplog.timestamp')
         except OSError:
             pass
-        self.cluster.stop()
+        cls.cluster.stop()
 
 
-class TestConnectorShardedMongosFailure(TestConnectorSharded):
+class TestConnectorSharded(ShardedConnectorTestCase):
+
+    def test_connector(self):
+        """Test whether the connector initiates properly
+        """
+        # Make sure get_mininum_mongodb_version returns the current version.
+        self.assertEqual(Version.from_client(self.cluster.client()),
+                         get_mininum_mongodb_version())
 
     def test_mongos_connection_failure(self):
         """Test that the connector handles temporary mongos failure"""
@@ -73,13 +89,30 @@ class TestConnectorShardedMongosFailure(TestConnectorSharded):
         assert_soon(lambda: len(self.dm._search()) == 3)
         client.close()
 
+    def test_connector_uri_options(self):
+        """Ensure the connector passes URI options to newly created clients.
+        """
+        expected_client = MongoClient(self.main_uri)
 
-class TestConnectorShardedAuth(TestConnectorSharded):
+        def assert_client_options(client):
+            self.assertEqual(client.read_preference,
+                             expected_client.read_preference)
+            self.assertEqual(client.read_preference,
+                             ReadPreference.PRIMARY_PREFERRED)
 
-    def setUp(self):
+        assert_client_options(self.connector.create_authed_client())
+        assert_client_options(self.connector.main_conn)
+        for oplog_thread in self.connector.shard_set.values():
+            assert_client_options(oplog_thread.primary_client)
+
+
+class TestConnectorShardedAuth(ShardedConnectorTestCase):
+
+    @classmethod
+    def setUp(cls):
         if not (db_user and db_password):
             raise SkipTest('Need to set a user/password to test this.')
-        super(TestConnectorShardedAuth, self).setUp()
+        super(TestConnectorShardedAuth, cls).setUpClass()
 
     def test_start_with_auth(self):
         # Insert some documents into the sharded cluster.  These